Works, but looks messy
This can turn logic into the equivalent of the 64 bit codes needed for the AI to use those logical rules as the genetic code of its mind.
The "programming language" is unlike any programming language you have used before, though if you have had practice coding in verilog, you can understand it.
The language paradigm is parallel imperative (not the normal sequential imperative that you are used to) bit-oriented. It is not object oriented, it is not functional (all variables are globals), and the only way to access ideas outside of the language (which consists of logical and "*", logical or "+", logical not "-", previous "p", and wait "wait") is through literal calls to the integer representing the relevant input neuron or output neuron.
The resulting file created from using this program as python3 Logic_to_AI_compiler_V3.py example_logic.txt example_logic.log
Still needs to be translated into binary, with "0" becoming a 0 binary digit rather than an 8 bit unicode character, but the .log file in the example is written in a way that you can visually check that the code is working.